requirements.txt
azure-functions
requests==2.19.1
O365==2.0.7